Spec configuration for backend api

+++ app/controllers/api/v1/users_controller.rb
@@ -4,11 +4,13 @@ module Api
       before_action :assign_params, only: :update
 
       def index
+        authorize(User)
         populator = UserPopulator.new(current_user: current_user, params: query_params)
         render_paginated_collection(populator.run)
       end
 
       def show
+        authorize(User)
         show_user_form(user_form)
       end
 
@@ -17,10 +19,12 @@ module Api
       end
 
       def update
+        authorize(User)
         update_user_form(user_form)
       end
 
       def destroy
+        authorize(User)
         destroy_user_form(user_form)
       end
